General cleanup and initial project setup.
