+++ app-cesards/app-init/build.gradle.kts
@@ -1,0 +1,10 @@
+apply(from = "$rootDir/gradle/android-library-common.gradle")
+
+plugins {
+  id("com.android.library")
+  kotlin("android")
+}
+
+dependencies {
+  implementation(project(":app-cesards:ui"))
+}
