+++ app-cesards/app-init/src/main/AndroidManifest.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.codeforvicotory.android.recipes">
+
+    <application android:theme="@style/Theme.MyRecipes">
+        <activity
+            android:name="org.codeforvictory.android.recipes.cesards.feature.ui.recipes.MainActivity"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.MyRecipes.NoActionBar">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
+
+</manifest>
