+++ app-cesards/app-init/src/main/res/values/strings.xml
