+++ app-cesards/ui-common/src/main/res/values/colors.xml
