+++ app-cesards/ui/src/main/java/org/codeforvictory/android/recipes/cesards/feature/ui/recipes/MainActivity.kt
@@ -1,0 +1,12 @@
+package org.codeforvictory.android.recipes.cesards.feature.ui.recipes
+
+import android.os.Bundle
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        Toast.makeText(this, "Hello world!", Toast.LENGTH_LONG).show()
+    }
+}
