+++ app-domain-common/model/build.gradle.kts
@@ -1,0 +1,6 @@
+apply(from = "$rootDir/gradle/android-library-common.gradle")
+
+plugins {
+  id("com.android.library")
+  kotlin("android")
+}
