--- app-domain/recipe/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="org.codeforvicotory.android.domain.recipe"/>
