+++ app-ui-common/recipe/src/main/AndroidManifest.xml
