--- app-ui-common/resources/src/main/java/org/codeforvicotory/android/recipes/ui/common/resources/Color.kt
@@ -1,13 +1,0 @@
-package org.codeforvicotory.android.recipes.ui.common.resources
-
-import androidx.compose.ui.graphics.Color
-
-private const val COLOR_PURPLE_200: Long = 0xFFBB86FC
-private const val COLOR_PURPLE_500: Long = 0xFF6200EE
-private const val COLOR_PURPLE_700: Long = 0xFF3700B3
-private const val COLOR_TEAL_200: Long = 0xFF03DAC5
-
-val purple200 = Color(COLOR_PURPLE_200)
-val purple500 = Color(COLOR_PURPLE_500)
-val purple700 = Color(COLOR_PURPLE_700)
-val teal200 = Color(COLOR_TEAL_200)
