--- app-ui-common/resources/src/main/java/org/codeforvicotory/android/recipes/ui/common/resources/Shape.kt
@@ -1,7 +1,0 @@
-package org.codeforvicotory.android.recipes.ui.common.resources
-
-/*val shapes = Shapes(
-  small = RoundedCornerShape(4.dp),
-  medium = RoundedCornerShape(4.dp),
-  large = RoundedCornerShape(0.dp)
-)*/
