--- app-ui-common/resources/src/main/java/org/codeforvicotory/android/recipes/ui/common/resources/Theme.kt
@@ -1,36 +1,0 @@
-package org.codeforvicotory.android.recipes.ui.common.resources
-
-/*private val DarkColorPalette = darkColors(
-  primary = purple200,
-  primaryVariant = purple700,
-  secondary = teal200
-)
-private val LightColorPalette = lightColors(
-  primary = purple500,
-  primaryVariant = purple700,
-  secondary = teal200
-  *//* Other default colors to override
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black,
-    *//*
-)
-
-@Composable
-fun RecipesMVIStudyGroupTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable() () -> Unit) {
-  val colors = if (darkTheme) {
-    DarkColorPalette
-  } else {
-    LightColorPalette
-  }
-
-  MaterialTheme(
-    colors = colors,
-    typography = typography,
-    shapes = shapes,
-    content = content
-  )
-}*/
