--- app-ui-common/resources/src/main/java/org/codeforvicotory/android/recipes/ui/common/resources/Type.kt
@@ -1,22 +1,0 @@
-package org.codeforvicotory.android.recipes.ui.common.resources
-
-// Set of Material typography styles to start with
-// val typography = Typography(
-  /*body1 = TextStyle(
-    fontFamily = FontFamily.Default,
-    fontWeight = FontWeight.Normal,
-    fontSize = 16.sp
-  )*/
-  /* Other default text styles to override
-    button = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.W500,
-        fontSize = 14.sp
-    ),
-    caption = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 12.sp
-    )
-    */
-// )
