--- app-ui/TODO.md
@@ -1,3 +1,0 @@
-# TODO
-
-Add your own module, representing the entry point of the app + anything on the UI really. You are in control.
