--- app-ui/cesards/src/main/java/org/codeforvicotory/android/recipes/ui/MainActivity.kt
@@ -1,10 +1,0 @@
-package org.codeforvicotory.android.recipes.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
