+++ gradle/android-library-common.gradle
@@ -16,4 +16,8 @@ android {
   kotlinOptions {
     jvmTarget = "1.8"
   }
+
+  kotlinOptions {
+    freeCompilerArgs = ["-Xallow-result-return-type"]
+  }
 }
